re refactoring
